tools/requirements.txt
pyserial>=3.5
pandas>=1.3.0
matplotlib>=3.4.0
